Corrected implementation of editing
This allows for the editing endpoint to work, but there is still an error that allows for a string to be passed in where a number is needed. I think this is a non-issue.

There is also currently not an error for if a field is not present in the database, but is trying to be added, but the field is not added to the database. This is not an issue as long as handled properly on frontend.

These are issues to be handled in bug fixes in the end